requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.4.0
pymc>=5.0.0
aesara>=0.2.0
arviz>=0.11.0
matplotlib>=3.4.0
seaborn>=0.11.0
statsmodels>=0.13.0
plotly>=5.3.0
tqdm>=4.66.1
joblib>=1.3.2
kaleido>=0.2.1
pmdarima==2.0.3
prophet==1.1.4 